pyproject.toml
[project]
name = "ai-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.170.0",
    "google-auth-oauthlib>=1.2.2",
    "pydantic>=2.11.5",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "python-dotenv[cli]>=1.1.0",
]
